requirements.txt
aiogram == 3.4.1
aiohttp~=3.9.3
SQLAlchemy~=2.0.27
python-dotenv~=1.0.1